+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.31.0] - 2024-08-21
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `/meals/create` page for creating new meals.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputList` component integrating `Autocomplete` from `MUI` library.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputImage` component for uploading images.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputSelect` component for selecting one of many options.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputCheckbox` component for checking option.
+
 ## [1.30.0] - 2024-08-21
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): Simplified recipe step structure to `string`.